requirements.txt
Flask
Flask-MySQL
Flask-SQLAlchemy
Flask-WTF
Flask-Login
mysql-connector-python
